feat: Implement HorizontalScrollBox as general component and enhance scrolling behavior
