+++ src/component/game/common/WinnerHistory/jsx/WinnerRectangle.jsx
@@ -1,48 +1,21 @@
-import React, { useEffect, useRef } from "react";
+import React from "react";
 import { WinnerBox } from "./WinnerBox";
 import style from "../style/WinnerHistory.module.css";
+import { HorizontalScrollBox } from "../../layout/jsx/HorizontalScrollBox";
 
 export const WinnerRectangle = ({ results }) => {
-  const winnerRectangleRef = useRef(null);
-
-  useEffect(() => {
-    if (winnerRectangleRef.current) {
-      winnerRectangleRef.current.scrollLeft =
-        winnerRectangleRef.current.scrollWidth;
-    }
-  }, [results]);
-
-  useEffect(() => {
-    const handleWheel = (event) => {
-      if (winnerRectangleRef.current) {
-        if (event.deltaY !== 0) {
-          winnerRectangleRef.current.scrollLeft += event.deltaY;
-        }
-      }
-    };
-
-    const currentRef = winnerRectangleRef.current;
-    if (currentRef) {
-      currentRef.addEventListener("wheel", handleWheel);
-    }
-
-    return () => {
-      if (currentRef) {
-        currentRef.removeEventListener("wheel", handleWheel);
-      }
-    };
-  }, []);
-
   return (
-    <div className={style.WinnerRectangle} ref={winnerRectangleRef}>
-      {results.map((result, index) => (
-        <WinnerBox
-          key={index}
-          winner={result.winner}
-          gameId={result.gameId}
-          roundId={result.roundId}
-        />
-      ))}
+    <div className={style.WinnerRectangle}>
+      <HorizontalScrollBox>
+        {results.map((result, index) => (
+          <WinnerBox
+            key={index}
+            winner={result.winner}
+            gameId={result.gameId}
+            roundId={result.roundId}
+          />
+        ))}
+      </HorizontalScrollBox>
     </div>
   );
 };
